Add test for ADC pins
Adds the zephyr core's test for the Analog to Digital Converter, which allows
viewing ADC pin voltage values over the zephyr serial console. Makes
corresponding changes to the STM32F405 Feather Express test profile to enable
the ADC.

In addition, this test adds a board overlay for the test STM32F405 feather,
enabling the specific channels required, matching other implementations of this
sample from the zephyr core. This channel information may be relocated in the
future, depending on the broader strategy for implementing board and feature
agnostic tests.