get Feather RP2350 SPI working
Turns out my config worked other than having `status = "okay";` in
the wrong node. Fixed that and now I can use the spi commands in
the shell. SCK, MOSI, and MISO definitely work. Not sure what's up
with CS. It's always low on the logic analyzer, which is good, but
I'm not sure if that's intentional or the result of the pin not
being configured at all.