Adafruit QT-Py ESP32S2 support
Adds support for the Adafruit QT-Py boards based off the ESP32-S2 (both base and
uFL antenna variant). Supports I2C and UART, tested with the corresponding
applications in this repo. Also supports SPI and the ADC, but will not work with
existing tests pending some required cross-board compatibility changes. Does not
support the blinky test as it does not have a simple LED, pending blinky
autodetection of a WS2812/Neopixel.